Removes buttons
